+++ web/specs/src/main/java/org/seedstack/seed/web/WebConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2013-2018, The SeedStack authors <http://seedstack.org>
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -227,8 +227,14 @@ public class WebConfig {
             return Collections.unmodifiableList(welcomeFiles);
         }
 
-        public void addWelcomeFile(String welcomeFile) {
+        public ServerConfig setWelcomeFiles(List<String> welcomeFiles) {
+            this.welcomeFiles = new ArrayList<>(welcomeFiles);
+            return this;
+        }
+
+        public ServerConfig addWelcomeFile(String welcomeFile) {
             this.welcomeFiles.add(welcomeFile);
+            return this;
         }
 
         @Config("sessions")