+++ web/undertow/src/main/java/org/seedstack/seed/undertow/internal/UndertowLauncher.java
@@ -1,11 +1,10 @@
 /*
- * Copyright © 2013-2018, The SeedStack authors <http://seedstack.org>
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-
 package org.seedstack.seed.undertow.internal;
 
 import io.nuun.kernel.api.Kernel;
@@ -35,7 +34,8 @@ import org.xnio.XnioWorker;
 
 public class UndertowLauncher implements SeedLauncher {
     private static final Logger LOGGER = LoggerFactory.getLogger(UndertowLauncher.class);
-    private final AtomicBoolean launched = new AtomicBoolean(false);
+    private static final AtomicBoolean firstRun = new AtomicBoolean(true);
+    private static final AtomicBoolean launched = new AtomicBoolean(false);
     private XnioWorker xnioWorker;
     private DeploymentManager deploymentManager;
     private Undertow undertow;
@@ -70,7 +70,6 @@ public class UndertowLauncher implements SeedLauncher {
         if (launched.get()) {
             LOGGER.info("Refreshing Web application");
             stopAll();
-            Seed.refresh();
             startAll();
         } else {
             throw SeedException.createNew(UndertowErrorCode.UNDERTOW_NOT_LAUNCHED);
@@ -85,6 +84,9 @@ public class UndertowLauncher implements SeedLauncher {
     }
 
     private void startAll() throws Exception {
+        if (!firstRun.compareAndSet(true, false)) {
+            Seed.refresh();
+        }
         Coffig baseConfiguration = Seed.baseConfiguration();
         createWorker(baseConfiguration);
         deploy(baseConfiguration);