+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 3.8.1 (xxxx-xx-xx)
-* [chg] Filter priorities updated: resource filter has -2000 priority, and Jersey filter has -1000 priority.
 
+* [new] Support for the `web.server.welcomeFiles` with Undertow. Default value is `index.html`.
+* [chg] Filter priorities updated: resource filter has -2000 priority, and Jersey filter has -1000 priority.
 
 # Version 3.8.0 (2018-11-27)